refactor: refactor job status types at collector page
